Fixed deprecation warning

+++ All in One/dependencies/openpiv_filters.py
@@ -215,8 +215,8 @@ def replace_nans_py(array, max_iter, tol, kernel_size=2, method='disk'):
 
       """
 
-    DTYPEf = np.float
-    DTYPEi = np.int
+    DTYPEf = float
+    DTYPEi = int
 
     filled = array.copy()
     n_dim = len(array.shape)
